pyproject.toml
[tool.black]
line-length = 99
target-version = ["py37", "py38", "py39", "py310", "py311"]
exclude = '''
/(
    \.git|.venv
)/
'''

[tool.isort]
profile = "black"
float_to_top = true
skip_glob = ".venv"

[tool.poetry]
name = "pubsub-emulator-messaging-setup"
version = "0.1.0dev0"
description = "CLI tool to setup topics and subscriptions on a PubSub emulator from a definition in a YAML file"
repository = "https://github.com/expobrain/pubsub-emulator-messaging-setup"
authors = ["Daniele Esposti <daniele.esposti@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pubsub_emulator_messaging_setup" }]

[tool.poetry.dependencies]
python = ">=3.7,<3.12"
click = ">=8.1"
pydantic = ">=1.10"
PyYAML = ">=6.0"
grpcio = ">=1.50"
google-cloud-pubsub = ">=2.13"
google-api-core = ">=2.10"
google-auth = ">=2.13"
loguru = ">=0.6"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.4"
black = "^22.1"
isort = "^5.10.1"
mypy = "^0.981"
pdbpp = "^0.10.3"
pytest = "^7.0.1"
pytest-cov = "^4.0.0"
flake8 = "^5.0.2"
flake8-datetime-utcnow-plugin = "^0.1.2"
pyupgrade = "^2.31.1"
flake8-bugbear = "^22.6.22"
pytest-clarity = "^1.0.1"
bandit = "^1.7.4"
types-PyYAML = "^6.0.12.1"
pytest-testmon = "^1.4.1"
pytest-timeout = "^2.1.0"
pytest-mock = "^3.10.0"

[tool.poetry.scripts]
pubsub_emu_setup = "pubsub_emulator_messaging_setup.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ".venv/"
check_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true

[[tool.mypy.overrides]]
module = ["google.auth.credentials", "grpc"]
ignore_missing_imports = true

[tool.pytest.ini_options]
timeout = 30
